Docs: 주석 추가(+ Calculation 디렉토리 분리)
